requirements.txt
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
outcome==1.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
selenium==4.4.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.11
wsproto==1.1.0
